+++ database/repositories/tyre_impression_repository.py
@@ -1,3 +1,4 @@
+from sqlalchemy.orm import joinedload
 from database.models import TyreImpression
 from database.repositories.base_repository import BaseRepository
 
@@ -6,5 +7,13 @@ class TyreImpressionRepository(BaseRepository[TyreImpression]):
 
     def __init__(self):
         super().__init__(TyreImpression)
+
+    def get_by_id(self, id_: int):
+        return (
+            self.db.query(TyreImpression)
+            .options(joinedload(TyreImpression.processing))
+            .filter(TyreImpression.id == id_)
+            .first()
+        )